v5.5 Test

+++ esp32p4/include/soc/include/soc/rtc_io_periph.h
@@ -13,6 +13,7 @@
 #if SOC_RTCIO_PIN_COUNT > 0
 #include "soc/rtc_io_channel.h"
 #endif
+
 #include "soc/io_mux_reg.h"
 
 #ifdef __cplusplus
@@ -62,7 +63,7 @@ extern const rtc_io_desc_t rtc_io_desc[SOC_RTCIO_PIN_COUNT];
  * This is an internal function of the driver, and is not usually useful
  * for external use.
  */
-extern const int8_t rtc_io_num_map[SOC_GPIO_PIN_COUNT];
+extern const int rtc_io_num_map[SOC_GPIO_PIN_COUNT];
 #endif //SOC_RTCIO_PIN_COUNT > 0
 
 #ifdef __cplusplus
